server/db.py
import os
import mysql.connector


def get_connection():
	return mysql.connector.connect(
		host=os.getenv("DB_HOST", "localhost"),
		port=int(os.getenv("DB_PORT", "3306")),
		user=os.getenv("DB_USER", "root"),
		password=os.getenv("DB_PASSWORD", ""),
		database=os.getenv("DB_NAME", "flashcards"),
	)
